feat(pulgin): add  localSearch plugin
