+++ packages/varlet-vite-plugins/src/index.ts
@@ -2,3 +2,4 @@ export * from './markdown.js'
 export * from './html.js'
 export * from './inlineCss.js'
 export * from './copy.js'
+export * from './localSearch.js'
